Assignment_03/requirements.dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.dev.txt requirements.dev.in
#
anyio==3.6.2
    # via jupyter-server
appdirs==1.4.4
    # via sqlfluff
argon2-cffi==21.3.0
    # via
    #   jupyter-server
    #   nbclassic
    #   notebook
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
arrow==1.2.3
    # via isoduration
asttokens==2.2.1
    # via stack-data
attrs==22.2.0
    # via
    #   jsonschema
    #   pytest
babel==2.11.0
    # via jupyterlab-server
backcall==0.2.0
    # via ipython
beautifulsoup4==4.11.1
    # via nbconvert
black==22.12.0
    # via -r requirements.dev.in
bleach==5.0.1
    # via nbconvert
build==0.10.0
    # via pip-tools
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via argon2-cffi-bindings
cfgv==3.3.1
    # via pre-commit
chardet==5.1.0
    # via
    #   diff-cover
    #   sqlfluff
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via
    #   black
    #   pip-tools
    #   sqlfluff
colorama==0.4.6
    # via sqlfluff
comm==0.1.2
    # via ipykernel
debugpy==1.6.5
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
detect-secrets==1.4.0
    # via -r requirements.dev.in
diff-cover==7.3.2
    # via sqlfluff
distlib==0.3.6
    # via virtualenv
entrypoints==0.4
    # via jupyter-client
exceptiongroup==1.1.0
    # via pytest
executing==1.2.0
    # via stack-data
fastjsonschema==2.16.2
    # via nbformat
filelock==3.9.0
    # via virtualenv
flake8==6.0.0
    # via -r requirements.dev.in
fqdn==1.5.1
    # via jsonschema
identify==2.5.13
    # via pre-commit
idna==3.4
    # via
    #   anyio
    #   jsonschema
    #   requests
iniconfig==2.0.0
    # via pytest
ipykernel==6.20.2
    # via
    #   nbclassic
    #   notebook
ipython==8.8.0
    # via
    #   ipykernel
    #   jupyterlab
ipython-genutils==0.2.0
    # via
    #   nbclassic
    #   notebook
isoduration==20.11.0
    # via jsonschema
isort==5.12.0
    # via -r requirements.dev.in
jedi==0.18.2
    # via ipython
jinja2==3.1.2
    # via
    #   diff-cover
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbclassic
    #   nbconvert
    #   notebook
    #   sqlfluff
json5==0.9.11
    # via jupyterlab-server
jsonpointer==2.3
    # via jsonschema
jsonschema[format-nongpl]==4.17.3
    # via
    #   jupyter-events
    #   jupyterlab-server
    #   nbformat
jupyter-client==7.4.9
    # via
    #   ipykernel
    #   jupyter-server
    #   nbclassic
    #   nbclient
    #   notebook
jupyter-core==5.1.3
    # via
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   nbclassic
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
jupyter-events==0.6.3
    # via jupyter-server
jupyter-server==2.1.0
    # via
    #   jupyterlab
    #   jupyterlab-server
    #   nbclassic
    #   notebook-shim
jupyter-server-terminals==0.4.4
    # via jupyter-server
jupyterlab==3.5.2
    # via -r requirements.dev.in
jupyterlab-pygments==0.2.2
    # via nbconvert
jupyterlab-server==2.19.0
    # via jupyterlab
markupsafe==2.1.2
    # via
    #   jinja2
    #   nbconvert
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
mccabe==0.7.0
    # via flake8
mistune==2.0.4
    # via nbconvert
mypy-extensions==0.4.3
    # via black
nbclassic==0.4.8
    # via
    #   jupyterlab
    #   notebook
nbclient==0.7.2
    # via nbconvert
nbconvert==7.2.8
    # via
    #   jupyter-server
    #   nbclassic
    #   notebook
nbformat==5.7.3
    # via
    #   jupyter-server
    #   nbclassic
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.6
    # via
    #   ipykernel
    #   jupyter-client
    #   nbclassic
    #   notebook
nodeenv==1.7.0
    # via pre-commit
nose==1.3.7
    # via -r requirements.dev.in
notebook==6.5.2
    # via jupyterlab
notebook-shim==0.2.2
    # via nbclassic
packaging==23.0
    # via
    #   build
    #   ipykernel
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
    #   pytest
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
pathspec==0.10.3
    # via
    #   black
    #   sqlfluff
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.12.1
    # via -r requirements.dev.in
platformdirs==2.6.2
    # via
    #   black
    #   jupyter-core
    #   virtualenv
pluggy==1.0.0
    # via
    #   diff-cover
    #   pytest
pre-commit==2.21.0
    # via -r requirements.dev.in
prometheus-client==0.15.0
    # via
    #   jupyter-server
    #   nbclassic
    #   notebook
prompt-toolkit==3.0.36
    # via ipython
psutil==5.9.4
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.2
    # via stack-data
pycodestyle==2.10.0
    # via flake8
pycparser==2.21
    # via cffi
pyflakes==3.0.1
    # via flake8
pygments==2.14.0
    # via
    #   diff-cover
    #   ipython
    #   nbconvert
pyproject-hooks==1.0.0
    # via build
pyrsistent==0.19.3
    # via jsonschema
pytest==7.2.1
    # via sqlfluff
python-dateutil==2.8.2
    # via
    #   arrow
    #   jupyter-client
python-json-logger==2.0.4
    # via jupyter-events
pytz==2022.7.1
    # via babel
pyyaml==6.0
    # via
    #   detect-secrets
    #   jupyter-events
    #   pre-commit
    #   sqlfluff
pyzmq==25.0.0
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   nbclassic
    #   notebook
regex==2022.10.31
    # via sqlfluff
requests==2.28.2
    # via
    #   detect-secrets
    #   jupyterlab-server
rfc3339-validator==0.1.4
    # via
    #   jsonschema
    #   jupyter-events
rfc3986-validator==0.1.1
    # via
    #   jsonschema
    #   jupyter-events
send2trash==1.8.0
    # via
    #   jupyter-server
    #   nbclassic
    #   notebook
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
    #   rfc3339-validator
sniffio==1.3.0
    # via anyio
soupsieve==2.3.2.post1
    # via beautifulsoup4
spark==0.2.1
    # via -r requirements.dev.in
sqlfluff==1.4.5
    # via -r requirements.dev.in
stack-data==0.6.2
    # via ipython
tblib==1.7.0
    # via sqlfluff
terminado==0.17.1
    # via
    #   jupyter-server
    #   jupyter-server-terminals
    #   nbclassic
    #   notebook
tinycss2==1.2.1
    # via nbconvert
toml==0.10.2
    # via sqlfluff
tomli==2.0.1
    # via
    #   black
    #   build
    #   jupyterlab
    #   pyproject-hooks
    #   pytest
tornado==6.2
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   nbclassic
    #   notebook
    #   terminado
tqdm==4.64.1
    # via sqlfluff
traitlets==5.8.1
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   jupyter-events
    #   jupyter-server
    #   matplotlib-inline
    #   nbclassic
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
typing-extensions==4.4.0
    # via sqlfluff
uri-template==1.2.0
    # via jsonschema
urllib3==1.26.14
    # via requests
virtualenv==20.17.1
    # via pre-commit
wcwidth==0.2.6
    # via prompt-toolkit
webcolors==1.12
    # via jsonschema
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.4.2
    # via jupyter-server
wheel==0.38.4
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
